+++ apps/playground/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @ai-gui/playground
 
+## 0.0.43
+
+### Patch Changes
+
+- Updated dependencies
+  - @ai-gui/core@0.30.0
+  - @ai-gui/devtools@0.30.0
+  - @ai-gui/plugin-artifact@0.30.0
+  - @ai-gui/plugin-citation@0.30.0
+  - @ai-gui/plugin-map@0.30.0
+  - @ai-gui/plugin-mermaid@0.30.0
+  - @ai-gui/plugin-molecule@0.30.0
+  - @ai-gui/plugin-solid@0.30.0
+  - @ai-gui/plugin-ui@0.30.0
+  - @ai-gui/react@0.30.0
+  - @ai-gui/vanilla@0.30.0
+  - @ai-gui/vue@0.30.0
+
 ## 0.0.42
 
 ### Patch Changes